Creates PlaceCard component, moves each component to separate file
