+++ src/components/place-card/place-card.jsx
@@ -1,0 +1,48 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const PlaceCard = (props) => {
+  const {placeName} = props;
+
+  return <article className="cities__place-card place-card">
+    <div className="place-card__mark">
+      <span>Premium</span>
+    </div>
+    <div className="cities__image-wrapper place-card__image-wrapper">
+      <a href="#">
+        <img className="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place image"/>
+      </a>
+    </div>
+    <div className="place-card__info">
+      <div className="place-card__price-wrapper">
+        <div className="place-card__price">
+          <b className="place-card__price-value">&euro;180</b>
+          <span className="place-card__price-text">&#47;&nbsp;night</span>
+        </div>
+        <button className="place-card__bookmark-button button" type="button">
+          <svg className="place-card__bookmark-icon" width="18" height="19">
+            <use xlinkHref="#icon-bookmark"></use>
+          </svg>
+          <span className="visually-hidden">To bookmarks</span>
+        </button>
+      </div>
+      <div className="place-card__rating rating">
+        <div className="place-card__stars rating__stars">
+          <span style={{width: `100%`}}></span>
+          <span className="visually-hidden">Rating</span>
+        </div>
+      </div>
+      <h2 className="place-card__name">
+        <a href="#">{placeName}</a>
+      </h2>
+      <p className="place-card__type">Apartment</p>
+    </div>
+  </article>;
+};
+
+
+PlaceCard.propTypes = {
+  placeName: PropTypes.string.isRequired
+};
+
+export default PlaceCard;
