+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from './components/app.jsx';
+import App from './components/app/app.jsx';
 
 const PLACE_NAMES = [`Wood and stone place`, `Canal View Prinsengracht`, `Nice, cozy, warm big bed apartment`, `Beautiful &amp; luxurious apartment at great location`];
